parse: just return the AST, not a channel
Our language is expression oriented, so each document will be a single
expression.  It doesn't really make sense then, to have a 'list' of
root nodes (i.e. a statement) which is what a channel implies (kind of).

This also implies we won't be doing incremental parsing.  We could later,
and it could be a neat goal if we want a really efficient LSP, but we're
gonna stash that idea for now.