parse: a document is just a single expression
but the parser still loops because sibling expressions are appended
into a single node.  This is scary because the current architecture
has a potential for infinite loops both in the lexer and parser.
We'll revisit and fix this in the future.  This commit leaves some
noise behind that makes debugging a bit easier - ultimately as we
harden this work we'll clean this stuff up